Feature ETP-4906: Extract shared NEO webhook fetch/unwrap plumbing
Checks: npm, testid, tests, regen, xml, pw-mocked, pw-integration
Hooks-Verified: v2 e79d7912 ad4c7f129780 85d43d534d9a
