+++ tools/app-shell/src/lib/rolesApi.js
@@ -1,78 +1,12 @@
-function detectBase() {
-  const path = window.location.pathname;
-  const webIdx = path.indexOf('/web/');
-  return webIdx !== -1 ? path.substring(0, webIdx) : (import.meta.env.VITE_API_BASE || '');
-}
+import { NEO_BASE, fetchNeoWebhookJson } from './neoWebhookClient.js';
 
-const BASE = detectBase();
-const NEO_BASE = `${BASE}/sws/neo`;
-
-function getToken() {
-  return localStorage.getItem('sf_auth_token');
-}
-
-/**
- * Shared fetch + response-unwrap logic for the `SFRolesOverview`/`SFSystemRoleTemplates`
- * webhook family — both are reached through the NEO pseudo-spec bridge (`NeoGoWebhookBridge`
- * in `com.etendoerp.go`, see its class javadoc) rather than the Webhooks module's
- * `/webhooks/*` dispatch, and both reproduce the exact same `{result: "<json-string>"}` /
- * `{error: "<message>"}` response envelope, so this one helper serves either. `webhookName`
- * is only used to make error messages point at the right backend class.
- *
- * Same fetch conventions as `lib/menuTree.js`'s `callMenuWebhook` — this must run as the
- * CURRENT logged-in user's own role (`sf_auth_token`), not an admin token, since the backend
- * itself decides admin/client-admin access (`NeoAccessHelper.isAdminOrClientAdmin`) from that
- * same role. A non-admin caller (or no role at all) gets `{ roles: [] }` back from the backend
- * — that is not an error, it's the documented "denied" shape — callers should render the
- * empty-state message for `roles.length === 0`, not treat it as a fetch failure.
- *
- * No `Content-Type` header: this is a GET with no body, and `application/json` isn't a
- * CORS-safelisted value — setting it unnecessarily triggers a preflight OPTIONS request (and
- * risks it failing) whenever `VITE_API_BASE` points at a different origin than the SPA.
- *
- * @param {string} path - the NEO pseudo-spec path segment (e.g. `'rolesoverview'`).
- * @param {string} webhookName - the backend webhook class name, for error messages only
- *   (e.g. `'SFRolesOverview'`).
- * @returns {Promise<object>} the unwrapped `{roles: [...]}` payload.
- */
-async function fetchNeoWebhookRoles(path, webhookName) {
-  const token = getToken();
-  const headers = {};
-  if (token) headers['Authorization'] = `Bearer ${token}`;
-
-  const res = await fetch(`${NEO_BASE}/${path}`, { headers });
-  const text = await res.text();
-  let data;
-  let parsed = true;
-  try { data = JSON.parse(text); } catch { data = text; parsed = false; }
-
-  if (!res.ok) {
-    throw new Error(data?.error || data?.message || `${webhookName} error: ${res.status}`);
-  }
-  // A 200 with a non-JSON body (e.g. a SPA-fallback index.html served when this endpoint isn't
-  // actually backed — no dev proxy, no backend, as in most E2E test environments) must be
-  // treated as a failure rather than silently resolving to an empty-but-valid roles list.
-  if (!parsed || typeof data !== 'object' || data === null) {
-    throw new Error(`${webhookName} returned a non-JSON response`);
-  }
-  if (data.error) {
-    throw new Error(data.error);
-  }
-  if (typeof data.result === 'string') {
-    try {
-      return JSON.parse(data.result);
-    } catch {
-      throw new Error(`${webhookName} returned an invalid result payload`);
-    }
-  }
-  if (data.result && typeof data.result === 'object' && !Array.isArray(data.result)) {
-    return data.result;
-  }
-  if (Array.isArray(data.roles)) {
-    return data;
-  }
-  throw new Error(`${webhookName} returned an unexpected shape`);
-}
+// A non-admin caller (or no role at all) gets `{ roles: [] }` back from the backend for both
+// webhooks below — that is not an error, it's the documented "denied" shape — callers should
+// render the empty-state message for `roles.length === 0`, not treat it as a fetch failure.
+// Shared fetch/parse/unwrap mechanics (base URL, token, `{result: "<json-string>"}` envelope)
+// live in `neoWebhookClient.js`'s `fetchNeoWebhookJson` — see that module for the full
+// rationale (no `Content-Type` header, fresh token per call, etc.).
+const rolesFallback = (data) => (Array.isArray(data.roles) ? data : null);
 
 /**
  * Fetches the GOClient roles overview from `GET /sws/neo/rolesoverview` (ETP-4513).
@@ -102,7 +36,7 @@ async function fetchNeoWebhookRoles(path, webhookName) {
  *   tier: string}>}>}>}
  */
 export async function fetchRolesOverview() {
-  return fetchNeoWebhookRoles('rolesoverview', 'SFRolesOverview');
+  return fetchNeoWebhookJson(`${NEO_BASE}/rolesoverview`, 'SFRolesOverview', rolesFallback);
 }
 
 /**
@@ -120,7 +54,8 @@ export async function fetchRolesOverview() {
  * instead, matching the ticket's target architecture: "no template role should be at client
  * level, only at system level."
  *
- * Same unwrap/error conventions as `fetchRolesOverview()` (shared via `fetchNeoWebhookRoles`).
+ * Same unwrap/error conventions as `fetchRolesOverview()` (shared via `neoWebhookClient.js`'s
+ * `fetchNeoWebhookJson`).
  * No `userCount`, no client-admin row in the response — see `SFSystemRoleTemplates.java`'s class
  * javadoc for why. Combine with `fetchRolesOverview()` when a caller also needs the tenant's
  * client-admin row (e.g. the Users grid's role filter/chips, which must still surface classic
@@ -130,5 +65,5 @@ export async function fetchRolesOverview() {
  *   windows: Array<{id: string, name: string, tier: string}>}>}>}
  */
 export async function fetchTemplateRoles() {
-  return fetchNeoWebhookRoles('systemroletemplates', 'SFSystemRoleTemplates');
+  return fetchNeoWebhookJson(`${NEO_BASE}/systemroletemplates`, 'SFSystemRoleTemplates', rolesFallback);
 }
